requirements.txt
absl_py==1.4.0
absl_py==1.3.0
matplotlib==3.3.1
matplotlib==3.5.1
ml_collections==0.1.1
nibabel==3.2.1
numpy==1.21.5
numpy==1.18.5
opencv_python==4.6.0.66
opencv_python_headless==4.6.0.66
Pillow==7.0.0
Pillow==8.2.0
Pillow==10.1.0
scikit_image==0.19.3
scikit_learn==0.24.2
scipy==1.4.1
scipy==1.7.3
seaborn==0.11.2
tensorflow==2.3.0
tensorflow==2.9.0
tensorflow_gan==2.0.0
torch==1.7.0
torchvision==0.8.0
tqdm==4.64.0
visdom==0.1.8.9
